Modify parameterizable class JpaRepository's parameter data type

+++ back-end/src/main/java/com/healthgenic/entities/user/dao/UserDaoInterface.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.healthgenic.entities.user.model.User;
 
-public interface UserDaoInterface extends JpaRepository<User, Long> {
+public interface UserDaoInterface extends JpaRepository<User, Integer> {
 
 }
